Add db interface
