+++ src/node/index.ts
@@ -2,7 +2,7 @@ import BN from "bn.js";
 import deepmerge from "deepmerge";
 
 import {BeaconChain} from "../chain";
-import {DB} from "../db";
+import {LevelDB} from "../db";
 import {Eth1Notifier} from "../eth1";
 import {P2PNetwork} from "../p2p";
 import {BeaconRPC} from "../rpc";
@@ -40,7 +40,7 @@ class BeaconNode {
     this.conf = deepmerge(defaultConf, opts);
 
     // this.logger ?
-    this.db = new DB(this.conf.db);
+    this.db = new LevelDB(this.conf.db);
     this.network = new P2PNetwork(this.conf.network);
     this.eth1 = new Eth1Notifier(this.conf.eth1);
     this.sync = new Sync(this.conf.sync, {
